Relocate GUI to grapics.ui
