+++ src/com/thunder_cut/grapics/ui/LeftPanel.java
@@ -3,7 +3,7 @@
  * Author : 김태건
  * Created Date : 2020-01-08
  */
-package GUI;
+package com.thunder_cut.grapics.ui;
 
 import javax.swing.*;
 import java.awt.*;
